Add verifyEmail/sendVerifyEmail to SessionAuthProvider and remove old AuthProvider
