--- frontend/src/components/auth/AuthProvider.tsx
@@ -1,110 +1,0 @@
-import { UserMe } from "@/api/users/userTypes.ts";
-import React, { createContext, useContext, useEffect, useState } from "react";
-import { ACCESS_TOKEN, REFRESH_TOKEN } from "@/constants.ts";
-import api from "@/api/apiClient.ts";
-
-interface AuthContextProps {
-  isAuthenticated: boolean;
-  user: UserMe | null;
-  login: (email: string, password: string) => Promise<void>;
-  register: (
-    firstName: string,
-    lastName: string,
-    email: string,
-    password: string,
-  ) => Promise<void>;
-  sendVerificationEmail: () => Promise<void>;
-  verifyEmail: (uid: string, token: string) => Promise<void>;
-  logout: () => void;
-  isLoading: boolean;
-}
-
-const AuthContext = createContext<AuthContextProps | undefined>(undefined);
-
-export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
-  children,
-}) => {
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-  const [user, setUser] = useState<UserMe | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  const login = async (email: string, password: string) => {
-
-  };
-
-  const logout = () => {
-
-  };
-
-  const register = async (
-    firstName: string,
-    lastName: string,
-    email: string,
-    password: string,
-  ) => {
-
-  };
-
-  const sendVerificationEmail = async () => {
-    await api.get("/users/send-email-verification/");
-  };
-
-  const verifyEmail = async (uid: string, token: string) => {
-    await api.post("/users/verify-email/", { uid, token });
-  }
-
-  useEffect(() => {
-    const interceptor = api.interceptors.response.use(
-      (response) => response,
-      async (error) => {
-        const originalRequest = error.config;
-
-        if (error.response?.status === 401 && !originalRequest._retry) {
-          originalRequest._retry = true;
-
-          try {
-            const newToken = await refreshAuthToken();
-            originalRequest.headers.Authorization = `Bearer ${newToken}`;
-            return api(originalRequest);
-          } catch(refreshError) {
-            logout();
-            return Promise.reject(refreshError);
-          }
-        }
-
-        return Promise.reject(error);
-      },
-    );
-
-    return () => {
-      api.interceptors.response.eject(interceptor);
-    };
-  }, []);
-
-  return (
-    <AuthContext.Provider
-      value={{
-        isAuthenticated,
-        user,
-        login,
-        register,
-        sendVerificationEmail,
-        verifyEmail,
-        logout,
-        isLoading,
-      }}
-    >
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-
-  if (!context) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-
-  return context;
-};
